\name{get_pptx_template}
\alias{get_pptx_template}
\title{Retrieves PowerPoint template file from the specified location.}
\usage{
get_pptx_template(pptx_dir,template_name)
}
\description{
Loads in a specified PowerPoint file from a specified locations. The slide creation functions will append their respective slides to this specified PowerPoint file.
}
\arguments{
\item{pptx_dir}{The directory that the PowerPoint template file is located in.}
\item{template_name}{The name of the PowerPoint template file.}
}
\examples{
"/Users/janedoe/Downloads","TemplatePPTX.pptx"
}
